diplomacy,Resources: Remove timebase-frequency in cpu anchor
After reviewing the RISC-V device-tree specification and RISC-V Linux
code it is clear that timebase-frequency should be per cpu node and not
in the anchor. It is a kernel bug that causes the kernel to panic when
there is no "timebase-frequency" property in "/cpus".
The problem should be addressed in the kernel, which I have done here
(and upstream seems to be working on it too, see the full commit message):

github.com/olajep/riscv-linux.git
    commit cd1f960fb8c2135472de00d6faf72141d576fbf6
    Author: Ola Jeppsson <ola.jeppsson@gmail.com>
    Date:   Thu Jan 10 11:33:17 2019 -0500

        RISC-V: Fix panic on missing "/cpus/timebase-frequency"