trace: Aggregator: Try to improve timing slack
Try to improve timing by optimizing buffer address calculation

Not sure if this really works...
Looking at the Vivado design timing report we're not among the top
offenders but aggregator shows up in the implementation timing report.
The timing constraints are met so it shouldn't be a problem, and maybe
Vivado doesn't care to optimize as long as timings are met????

We can probably decrease the width of trace_offset though. No need for
4GB size buffers ...

Signed-off-by: Ola Jeppsson <ola.jeppsson@gmail.com>