fpga: pardcore: dts: Let Linux handle UART0
Let Linux handle UART0 and don't trap to SBI for writing to the console.
Also rename UART1 --> UART0, UART2 --> UART1 ... to reflect the change.

The Linux RISC-V port doesn't really support earlycon on ttyUL0 but it
falls back to hci0 (SBI) which will do for now.

[ It depends on CONFIG_FIX_EARLYCON_MEM which needs fixmap.h. This is
currently missing in the RISC-V port (in 4.18) ]