sifive-blocks: Sync UART with upstream
LvNA BBL was hard-coded to use the uartlite console driver, instead of
deriving this data from the device-tree. I've reverted this behaviour in
the BBL code.
As a consequence of the above the LvNA UART's register layout had to be
modified so that the BBL uartlite console driver would work with the
device.

But now that we pull the right driver from the device-tree we also need
to restore the UART's register layout.

Sync up until this commit:
7b7fa7832e0dd102c7da2db740eaef2aebafbe35
https://github.com/sifive/sifive-blocks/commit/7b7fa7832e0dd102c7da2db740eaef2aebafbe35

... but retain the modifications that were made to the UART in the LvNA
codebase. To do that this commit adds the following parameters to
UARTParams:

withRx:         If false pull the RX pin up.
withInterrupts: Enable/disable interrupts
withPrinter:    If true print serial output to the file serial${addr}
withFast:       If true set the divider to 1