trace: Implement double buffering
Add support to switch to the next buffer when the current trace buffer
becomes full. When a buffer gets full the irq line goes high.
It is then the job for the driver to update the buffer address and clear
the buffer full flag of the full buffer.

TODO: Still need to implement support for the driver clearing the
bufX_full flags.