trace: TraceLogic: Fix out of order edge-case
When a TimeStamp trace is generated one clock after a normal trace, it
will be injected to the trace stream before the normal trace.
This is due to the fact that we give higher priority to the TimeStamp
trace, and because the buffer queue has a one clock cycle delay from
input to output.

This fix does two things:
1. Remove the one clock-cycle delay from the normal trace buffer
('fifo'), by setting flow=true.

2. Turn timestamp_inject into a wire.
This doesn't help fixing the bug but will make the TimeStamp traces be
even 2^18 so they're easier to interpret.

Not verified, but I can't reproduce this with this change. And it makes
sense logically.