trace: TraceLogic: Implement MSBPC trace
We need the full PC to reconstruct the call graph.
Since we can't fit a 64 bit PC into one trace we have to split it into
two traces. This is wasteful since we're using 128 bits, so only inject
a MSBPC trace if the upper bits changed from the previous MSBPC trace.