fpga: pardcore: dts: Fix UART interrupt numbers
Change was introduced as a side effect in this commit:

  commit ca2bf28920c5f7ac381e6257bdb48e5591254a4a
  Author: Ola Jeppsson <ola.jeppsson@gmail.com>
  Date:   Wed Jul 17 15:03:59 2019 -0400

      trace: Share one TraceCtrl between all all RocketTiles

The code that wires up the trace control module was moved to its own
trait, HasTraceCtrl (later renamed to CanHaveTraceCtrl). But due to
how Scala inheritance works, the interrupt line is now connected to the
interrupt bus at a later stage.

I don't see an easy way to change the behavior back so instead just
adjust the devicetree.